requirements.txt
Flask==2.2.2
flask_cors==3.0.10
flask_pymongo==2.3.0
PyMuPDF==1.22.3
Pillow==9.5.0
bcrypt==3.2.0
